requirements.txt
smbus2==0.4.3
websocket-client==1.8.0
